backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
PyMySQL==1.1.1
python-dotenv==1.0.1
flask-cors==4.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.30